refactor: comprehensive codebase cleanup and decoupling (v2)
- Decouple large `use-conversation-store.ts` into smaller modules.
- Decouple large `use-chat.ts` using `useChatUser`, `useChatPersistedSettings`, and `useChatActions`.
- Centralize `charsPerToken` constant and apply DRY in token estimation.
- Improve type safety using `satisfies` operator and explicit types.
- Fix functional regression in conversation creation logic.
- Remove redundant comments and improve naming throughout.
- Add `VITE_APP_URL` to environment validation.
